refactor: organize hooks
