+++ frontend/src/components/general/cards/SensorCard.tsx
@@ -2,7 +2,7 @@ import { Sensor } from '@green-ecolution/backend-client'
 import { format, formatDistanceToNow } from 'date-fns'
 import React from 'react'
 import Pill from '../Pill'
-import { getSensorStatusDetails } from '@/hooks/useDetailsForSensorStatus'
+import { getSensorStatusDetails } from '@/hooks/details/useDetailsForSensorStatus'
 import { useQuery } from '@tanstack/react-query'
 import { treeSensorIdQuery } from '@/api/queries'
 import { Link } from '@tanstack/react-router'
