+++ frontend/src/components/sensor/SensorDashboard.tsx
@@ -4,7 +4,7 @@ import EntitiesStatusCard from '@/components/general/cards/EntitiesStatusCard'
 import GeneralStatusCard from '@/components/general/cards/GeneralStatusCard'
 import BackLink from '@/components/general/links/BackLink'
 import GeneralLink from '../general/links/GeneralLink'
-import { getSensorStatusDetails } from '@/hooks/useDetailsForSensorStatus'
+import { getSensorStatusDetails } from '@/hooks/details/useDetailsForSensorStatus'
 import { format, formatDistanceToNow } from 'date-fns'
 import { de } from 'date-fns/locale'
 import DetailedList from '../general/DetailedList'
