+++ frontend/src/components/treecluster/TreeClusterDashboard.tsx
@@ -4,7 +4,7 @@ import TreeCard from '@/components/general/cards/TreeCard'
 import BackLink from '@/components/general/links/BackLink'
 import ButtonLink from '@/components/general/links/ButtonLink'
 import { Pencil } from 'lucide-react'
-import { getWateringStatusDetails } from '@/hooks/useDetailsForWateringStatus'
+import { getWateringStatusDetails } from '@/hooks/details/useDetailsForWateringStatus'
 import { Link } from '@tanstack/react-router'
 import GeneralLink from '../general/links/GeneralLink'
 import { TriangleAlert } from 'lucide-react'
