+++ frontend/src/components/vehicle/VehicleDashboard.tsx
@@ -2,10 +2,10 @@ import BackLink from '../general/links/BackLink'
 import Pill from '../general/Pill'
 import ButtonLink from '../general/links/ButtonLink'
 import { Pencil } from 'lucide-react'
-import { getVehicleStatusDetails } from '@/hooks/useDetailsForVehicleStatus'
+import { getVehicleStatusDetails } from '@/hooks/details/useDetailsForVehicleStatus'
 import GeneralLink from '../general/links/GeneralLink'
 import { VehicleStatus } from '@green-ecolution/backend-client'
-import { getVehicleType } from '@/hooks/useDetailsForVehicleType'
+import { getVehicleType } from '@/hooks/details/useDetailsForVehicleType'
 import DetailedList from '../general/DetailedList'
 import { useSuspenseQuery } from '@tanstack/react-query'
 import { vehicleIdQuery } from '@/api/queries'
