+++ frontend/src/components/watering-plan/TabGeneralData.tsx
@@ -6,7 +6,7 @@ import DetailedList from '../general/DetailedList'
 import { format, formatDuration, intervalToDuration } from 'date-fns'
 import GeneralStatusCard from '../general/cards/GeneralStatusCard'
 import EntitiesStatusCard from '../general/cards/EntitiesStatusCard'
-import { getWateringPlanStatusDetails } from '@/hooks/useDetailsForWateringPlanStatus'
+import { getWateringPlanStatusDetails } from '@/hooks/details/useDetailsForWateringPlanStatus'
 import { useSuspenseQuery } from '@tanstack/react-query'
 import { userQuery } from '@/api/queries'
 import { de } from 'date-fns/locale'
