+++ frontend/src/components/watering-plan/WateringPlanUpdate.tsx
@@ -8,7 +8,7 @@ import { useNavigate } from '@tanstack/react-router'
 import { Route } from '@/routes'
 import useStore from '@/store/store'
 import GeneralLink from '../general/links/GeneralLink'
-import { showWateringPlanStatusButton } from '@/hooks/useDetailsForWateringPlanStatus'
+import { showWateringPlanStatusButton } from '@/hooks/details/useDetailsForWateringPlanStatus'
 import LoadingInfo from '../general/error/LoadingInfo'
 import { Suspense } from 'react'
 import DeleteSection from '../treecluster/DeleteSection'
