+++ frontend/src/hooks/details/useDetailsForDrivingLicense.ts
