+++ frontend/src/hooks/details/useDetailsForSensorStatus.ts
