+++ frontend/src/hooks/details/useDetailsForSoilCondition.ts
