+++ frontend/src/hooks/details/useDetailsForUser.ts
