+++ frontend/src/hooks/details/useDetailsForVehicleStatus.ts
