+++ frontend/src/hooks/details/useDetailsForVehicleType.ts
