+++ frontend/src/hooks/details/useDetailsForWateringPlanStatus.ts
