+++ frontend/src/hooks/details/useDetailsForWateringStatus.ts
