+++ frontend/src/routes/_protected/map/index.tsx
@@ -10,7 +10,7 @@ import { treeQuery } from '@/api/queries'
 import { useRef, useState } from 'react'
 import Dialog from '@/components/general/filter/Dialog'
 import StatusFieldset from '@/components/general/filter/fieldsets/StatusFieldset'
-import { getWateringStatusDetails } from '@/hooks/useDetailsForWateringStatus'
+import { getWateringStatusDetails } from '@/hooks/details/useDetailsForWateringStatus'
 import useFilter from '@/hooks/useFilter'
 import FilterProvider, { Filters } from '@/context/FilterContext'
 import { z } from 'zod'
